Create script to scrape the parents of congressmen
The script will fetch the data from the congress website and then it
will create a file in data/congressperson_relatives.xz with 3 columns,
namely 'congressperson_id', 'relationship' and 'relative_name'. This
file will contain the names of the parent(s) of the congressmen.
The script will also create another file called
data/congressperson_relatives_raw.xz that will contain the original
string obtained from the site in case the script could not properly
split the father's and mother's name.