Add lodging expense analysis based on simple mean/std comparisons
This should be considered as an initial approach to #26. Of course
one may argue that there isn't any official data being scraped and
considered yet, but this mean/std approach may prove to be useful
for further analysis, with or without external data.

This version uses the recently added reimbursements dataset #140.