Upgrade auto-conversion hook to remove deprecation
The current hook proposed in the best practices section of
CONTRIBUTING.md to auto-convert notebooks to .py and .html formats
has a slightly minor problem: it uses a deprecated command, as stated
in the stdout every time the hook gets called:

```
[TerminalIPythonApp] WARNING | Subcommand `ipython nbconvert` is
deprecated and will be removed in future versions.
[TerminalIPythonApp] WARNING | You likely want to use `jupyter
nbconvert` in the future
```

This commit fix the recommendation, orientating users to use the new
approach (jupyter nbconvert) instead of the old one. There should not
be any side effects.