improvements in fetching CNPJs info
brought back parallel fetching with the abillity to use http proxies. Now each thread chooses in a random way one proxy or none (use the local IP) to make the request.

The numer of threads and a proxy list can be passed as arguments to the script as:
` python ./src/fetch_cnpj_info.py ./data/2016-11-19-current-year.xz -p 177.67.84.135:8080 177.67.82.80:8080 177.67.82.80:3128 179.185.54.114:8080 -t 20`