Add filter by reimbursement number (document_id)
After the last architecture change we are able to store and list
reimbursements that were not properly reimbursed: expenses paid
directly by the Chamber of Deputies and whose values are deducted from
each congresspeople quota (without actually reimbursing them, since they
have not properly paid for this expense).

In other words, instead of congressperson Jane Doe paying company X and
claiming the reimbursement (presenting the receipt) to the Chamber of
Deputies, the Chamber directly pays company X (that is the case of
flight tickets and postal services, for example).

Thus, these reimbursements have no `document_id` and no receipt, so it might
be useful to filter them out in some use cases.