Run Ansible locally (with local SSH keys)
After feedback from the community (thanks @willianpaixao and
@filipecifali) Ansible runs locally now (and not inside Docker anymore).

This change also addresses the issue of creating a one-time only SSH
key: now the process uses the host machine's SSH private key, and only
requires the user to inform (as an environment variable) the name the
public part of this key holds in DigitalOcean.

It also uses Pipenv to be more explicit and restrict about the Python 2
requirement: by using it to run the commands it ensures the Python
version specified in the Pipfile.