Add HTTPError handling when Reimbursement URL doesn't exist
The issue happened when running Rosie on 2020-01-01. By the way the adapter.py works, it tries to fetch reimbursement files up to the year in which the application is running .e.g. 2020. Since there is no such URL (at the time of this commit), the whole process fails due to an HTTPError 404. 

By adding this handling, the reimbursements that cannot be found are skipped from the download process and the application runs without any issues.